Enqueue admin notices scripts and styles on separate methods
